Use Folia scheduler logic for tick loop
I noticed that the jitter on the tick start was currently
+-5ms, and up to 10ms - which is pretty terrible. The
Thread.sleep() call appears to be a big culprit.

The sleep logic now uses LockSupport park(), which allows
ns precision and should help with sleep duration. Additionally,
I've removed catchup to prevent random server speedups on
lag spikes.

The logic to determine the next tick start is copied directly
from Folia.

The jitter is now reduced to <1ms, which is likely the best
that we can get out of this.

Additionally, raise the tick sampling for /tps from 1s to every
tick. This should make the command more accurate, as it is now
representative of all ticks - which is what it is expected to be.

Add 5s, and 15s tick averages as well.