Improve response times of default networking code
The reader thread does not need to be sleeping at all, it can
just block on the next input read. Any socket close activity
will interrupt the read, which is already handled by the packet
read function. As a result, this eliminates extra input latency
imposed by the reader thread.

The writer thread does not wake up when a packet is ready to
write, which can add up to 100ms of write delay. Additionally,
the packets are only flushed after the sleep delay which doesn't
help. Rewrite the sleep logic so that any packet writes into the
writer will wake it from sleep and move the flush() logic to occur
directly after draining the packet queue.